requirements.txt
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
cryptography==46.0.7
dotenv-python==0.0.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
mcp==1.27.0
pycparser==3.0
pydantic==2.13.0
pydantic-settings==2.13.1
pydantic_core==2.46.0
PyJWT==2.12.1
python-dotenv==1.2.2
python-multipart==0.0.26
pywin32==311
referencing==0.37.0
requests==2.33.1
rpds-py==0.30.0
sse-starlette==3.3.4
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.44.0
